# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: SoulAuth
  version: "0.1.0"
  description: >-
    Actor-native identity and authentication infrastructure.
    Endpoint surface only; payload schemas are not yet frozen.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: "{APP_URL}"
    description: >-
      The public address, set by APP_URL. It is also the OIDC issuer — the two
      must match character for character.
    variables:
      APP_URL:
        default: http://localhost:8080
components:
...
  schemas:
    Error:
      type: object
      required:
      - error
      - message
      additionalProperties: true
      description: 'The single error envelope used across the whole API. `error` is a **stable machine-readable
        code** — branch on it. `message` is prose: its wording may improve at any time and must not be
        used for decisions. Individual errors carry documented additional fields.'
      properties:
        error:
          type: string
          description: Stable error code, snake_case.
          enum:
          - internal_error
          - invalid_credentials
          - email_not_verified
          - invalid_token
          - user_not_found
          - email_exists
          - username_exists
          - oauth_error
          - password_already_set
          - invalid_user_id
          - not_found
          - validation_error
          - permission_denied
          - insufficient_permissions
          - missing_permission
          - account_suspended
          - account_inactive
          - account_deleted
          - account_locked
          - rate_limit_exceeded
          - forbidden
          - bad_request
          - not_configured
          - unauthorized
          - service_unavailable
        message:
          type: string
          description: Human-readable explanation. Not part of the contract.
        required_permission:
          type: string
          description: Carried by `missing_permission` only — which permission is missing.
        locked_until_seconds:
          type: integer
          description: Carried by `account_locked` only — seconds until the lock clears.
...
    LockoutCheckResult:
      properties:
        is_locked:
          type: boolean
        lockout_type:
          type: object
        message:
          type: string
        remaining_attempts:
          type: integer
        remaining_lockout_seconds:
          type: integer
      required:
      - is_locked
      - remaining_attempts
      - message
      type: object
...
    UnlockRequest:
      properties:
        identifier:
          type: string
        scope:
          type: object
      required:
      - scope
      - identifier
      type: object
    UnlockResponse:
      properties:
        unlocked:
          type: boolean
      required:
      - unlocked
      type: object
...
  securitySchemes:
    # 三种，不是一种。
    #
    # 把它们混成一个 `bearerAuth` 会产生具体的误导：读契约的人会以为
    # /api/oidc/userinfo 收的是登录拿到的那枚会话令牌 —— 它收的是 OIDC
    # 访问令牌，两者由不同端点签发、生命周期也不同。
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Session token, from POST /api/auth/login (or POST
        /api/actors/authenticate for an AI actor). Used by every /api business
        endpoint.
    oidcAccessToken:
      type: http
      scheme: bearer
      description: >-
        OIDC access token, from POST /api/oidc/token. Used by
        /api/oidc/userinfo **only**. Not interchangeable with the session token
        above.
    browserSession:
      type: apiKey
      in: cookie
      name: soulauth_session
      description: >-
        Browser session cookie. /api/oidc/authorize uses it to decide whether
        this browser is logged in; its absence is a redirect to the login page,
        not an error.
...
paths:
...
  /api/security/lockout:
    get:
      tags:
      - Security
      operationId: security_get_lockout_status
      description: >-
        Current lockout records, per account and per IP.
      description_zh: >-
        返回当前的锁定记录，按账号与按 IP 两个维度。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:security.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LockoutCheckResult'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: identifier
        in: query
        required: true
        schema:
          type: string
      - name: scope
        in: query
        required: true
        schema:
          type: object
  /api/security/unlock:
    post:
      tags:
      - Security
      operationId: security_unlock
      description: >-
        Clear a lockout. Idempotent: unlocking something that is not locked returns `false`
        rather than erroring. Both lock and unlock are audited.
      description_zh: >-
        解除锁定。幂等：对未被锁定的对象解锁返回 `false` 而不是报错。上锁与解锁都会写审计。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:security.write
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnlockResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockRequest'
...